test: isolate behavior test state
