+++ tests/behavior/steps/cli_options_steps.py
@@ -35,7 +35,11 @@ def test_cli_primus_start(bdd_context):
     pass
 
 
-@when(parsers.re(r'I run `autoresearch search "(?P<query>.+)" --loops (?P<loops>\d+) --token-budget (?P<budget>\d+) --no-ontology-reasoning'))
+@when(
+    parsers.parse(
+        'I run `autoresearch search "{query}" --loops {loops:d} --token-budget {budget:d} --no-ontology-reasoning`'
+    )
+)
 def run_with_budget(query, loops, budget, monkeypatch, cli_runner, bdd_context):
     ConfigLoader.reset_instance()
     def mock_run_query(q, cfg, callbacks=None):
@@ -80,13 +84,16 @@ def check_agents_config(bdd_context, agents):
     assert bdd_context["result"].exit_code == 0
 
 
-@when(parsers.parse('I run `autoresearch search --parallel --agent-groups "{g1}" "{g2}" "{query}"'))
+@when(
+    parsers.re(
+        r'^I run `autoresearch search --parallel --agent-groups "(?P<g1>.+)" "(?P<g2>.+)" "(?P<query>.+)"`$'
+    )
+)
 def run_parallel_cli(g1, g2, query, monkeypatch, cli_runner, bdd_context):
     ConfigLoader.reset_instance()
     groups = [[a.strip() for a in g1.split(",")], [a.strip() for a in g2.split(",")]]
 
     def mock_parallel(q, cfg, agent_groups):
-        bdd_context["groups"] = agent_groups
         return QueryResponse(answer="ok", citations=[], reasoning=[], metrics={})
 
     monkeypatch.setattr(Orchestrator, "run_parallel_query", mock_parallel)
@@ -96,6 +103,7 @@ def run_parallel_cli(g1, g2, query, monkeypatch, cli_runner, bdd_context):
         ["search", "--parallel", "--agent-groups", g1, "--agent-groups", g2, query],
     )
     bdd_context["result"] = result
+    bdd_context["groups"] = groups
 
 
 @when(parsers.parse('I run `autoresearch search "{query}" --reasoning-mode {mode}`'))
